# swagger.yaml
openapi: 3.0.0
info:
  title: Your API Documentation
  version: 1.0.0
  description: API documentation for your Express.js application
paths:
...
  /api/v1/items:
    get:
      summary: Get all items
      description: Retrieve all items with pagination and search
      tags:
        - Items
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search query for item name
        - in: query
          name: page
          schema:
            type: integer
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
          description: Number of items per page
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        price:
                          type: number
                        description:
                          type: string
                        quantity:
                          type: number
                        image:
                          type: string
                        owner:
                          type: string
                          format: ObjectId
                  currentPage:
                    type: integer
                  totalPages:
                    type: integer
                  totalCount:
                    type: integer
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    post:
      summary: Add an item
      description: Add a new item
      tags:
        - Items
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                quantity:
                  type: number
                image:
                  type: string
                owner:
                  type: string
                  format: ObjectId
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  price:
                    type: number
                  description:
                    type: string
                  quantity:
                    type: number
                  image:
                    type: string
                  owner:
                    type: string
                    format: ObjectId
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  "/api/v1/items/{id}":
    get:
      summary: Get item by ID
      description: Retrieve an item by its ID
      tags:
        - Items
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Item ID
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  price:
                    type: number
                  description:
                    type: string
                  quantity:
                    type: number
                  image:
                    type: string
                  owner:
                    type: string
                    format: ObjectId
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    put:
      summary: Update item by ID
      description: Update an item by its ID
      tags:
        - Items
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Item ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                description:
                  type: string
                quantity:
                  type: number
                image:
                  type: string
                owner:
                  type: string
                  format: ObjectId
      responses:
        "203":
          description: Item updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  price:
                    type: number
                  description:
                    type: string
                  quantity:
                    type: number
                  image:
                    type: string
                  owner:
                    type: string
                    format: ObjectId
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      summary: Delete item by ID
      description: Delete an item by its ID
      tags:
        - Items
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: Item ID
      responses:
        "204":
          description: Item deleted successfully
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...